Rename otherCurrency to exchanges
